Navigation using touch gestures
It is now possible to use touch gestures to navigate. I've used
hammer.js to handle the events, but it is not as smooth as I thought.

Sometimes the device's natural content scrolling behavior is apparently
 "cancelled" by hammer. That's not good.

Until I make it work decently, this feature is disabled by default
(using the brand new Waker.config props).