Starting to work on Waker custom tags.
The idea of Waker custom tags is to provide an easy way to create
magazine-like stuff, such as the end mark at the end of each story.

It's much easier to use something like <w:end /> than a huge block of
HTML and CSS. This commit has two dummy tags: <w:end> and <w:qr> (for
QR code).